--- app/src/main/java/com/paulik/popularlibraries/domain/entity/ImageEntity.kt
@@ -1,8 +1,0 @@
-package com.paulik.popularlibraries.domain.entity
-
-import java.util.*
-
-data class ImageEntity(
-    val id: String = UUID.randomUUID().toString(),
-    val image: String
-)
