+++ app/src/main/java/com/paulik/popularlibraries/ui/users/UserGitHubMvpActivity.kt
@@ -3,12 +3,14 @@ package com.paulik.popularlibraries.ui.users
 import com.github.terrakok.cicerone.androidx.AppNavigator
 import com.paulik.popularlibraries.App
 import com.paulik.popularlibraries.R
+import com.paulik.popularlibraries.domain.UsersGitHubMvpView
+import com.paulik.popularlibraries.domain.entity.UsersGitHubEntity
 import com.paulik.popularlibraries.ui.users.base.BackButtonListener
 import com.paulik.popularlibraries.ui.users.details.DetailsUserGitHubFragment
 import moxy.MvpAppCompatActivity
 import moxy.ktx.moxyPresenter
 
-class UserGitHubMvpActivity : MvpAppCompatActivity(R.layout.activity_users),
+class UserGitHubMvpActivity : MvpAppCompatActivity(R.layout.activity_users), UsersGitHubMvpView,
     UsersGitHubMvpFragment.Controller {
 
     private val navigator = AppNavigator(this, R.id.container)
@@ -38,7 +40,11 @@ class UserGitHubMvpActivity : MvpAppCompatActivity(R.layout.activity_users),
         presenter.backPressed()
     }
 
-    fun showUser(user: String) {
+    override fun updateList(users: List<UsersGitHubEntity>) {
+        // TODO("Not yet implemented")
+    }
+
+    override fun showUser(user: String) {
         supportFragmentManager
             .beginTransaction()
             .replace(
