+++ app/src/main/res/layout/fragment_convertor_image.xml
@@ -13,15 +13,14 @@
         android:layout_weight="1"
         android:orientation="vertical">
 
-        <EditText
-            android:id="@+id/name_image_edit_text"
+        <TextView
+            android:id="@+id/name_image_text_view"
             android:layout_width="match_parent"
-            android:layout_height="58dp"
+            android:layout_height="78dp"
             android:layout_marginTop="5dp"
             android:layout_marginBottom="5dp"
             android:gravity="center_horizontal"
             android:hint="Имя файла"
-            android:inputType="textMultiLine"
             android:textColor="@color/black"
             android:textSize="@dimen/default_text" />
